requirements.txt
pdfkit
streamlit
streamlit-option-menu
langchain
streamlit-audiorecorder
